06/14/2019 | Added favicons, done refactoring.
