+++ public/browserconfig.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<browserconfig>
+    <msapplication>
+        <tile>
+            <square70x70logo src="./images/ms-icon-70x70.png"/>
+            <square150x150logo src="./images/ms-icon-150x150.png"/>
+            <square310x310logo src="./images/ms-icon-310x310.png"/>
+            <TileColor>#2196f3</TileColor>
+        </tile>
+    </msapplication>
+</browserconfig>
