+++ public/index.html
@@ -17,8 +17,8 @@
     />
 
     <meta name="msapplication-TileColor" content="#2196f3" />
-    <meta name="msapplication-TileImage" content="" />
-    <meta name="theme-color" content="#2196f3" />
+    <meta name="msapplication-TileImage" content="$%PUBLIC_URL%/images/ms-icon-144x144.png" />
+    <meta name="theme-color" content="#212121" />
     <meta charset="utf-8" />
   
     <!--
@@ -35,21 +35,21 @@
       Setting up all the icons that will be used to support different 
       devices and formats.
     -->
-    <link rel="apple-touch-icon" sizes="57x57" href="">
-    <link rel="apple-touch-icon" sizes="60x60" href="">
-    <link rel="apple-touch-icon" sizes="72x72" href="">
-    <link rel="apple-touch-icon" sizes="76x76" href="">
-    <link rel="apple-touch-icon" sizes="114x114" href="">
-    <link rel="apple-touch-icon" sizes="120x120" href="">
-    <link rel="apple-touch-icon" sizes="144x144" href="">
-    <link rel="apple-touch-icon" sizes="152x152" href="">
-    <link rel="apple-touch-icon" sizes="180x180" href="">
+    <link rel="apple-touch-icon" sizes="57x57" href="%PUBLIC_URL%/images/apple-icon-57x57.png">
+    <link rel="apple-touch-icon" sizes="60x60" href="%PUBLIC_URL%/images/apple-icon-60x60.png">
+    <link rel="apple-touch-icon" sizes="72x72" href="%PUBLIC_URL%/images/apple-icon-72x72.png">
+    <link rel="apple-touch-icon" sizes="76x76" href="%PUBLIC_URL%/images/apple-icon-76x76.png">
+    <link rel="apple-touch-icon" sizes="114x114" href="%PUBLIC_URL%/images/apple-icon-114x114.png">
+    <link rel="apple-touch-icon" sizes="120x120" href="%PUBLIC_URL%/images/apple-icon-120x120.png">
+    <link rel="apple-touch-icon" sizes="144x144" href="%PUBLIC_URL%/images/apple-icon-144x144.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="%PUBLIC_URL%/images/apple-icon-152x152.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="%PUBLIC_URL%/images/apple-icon-180x180.png">
 
-    <link rel="icon" type="image/png" sizes="192x192"  href="">
+    <link rel="icon" type="image/png" sizes="192x192"  href="%PUBLIC_URL%/images/android-icon-192x192.png">
 
-    <link rel="icon" type="image/png" sizes="32x32" href="">
-    <link rel="icon" type="image/png" sizes="96x96" href="">
-    <link rel="icon" type="image/png" sizes="16x16" href="">
+    <link rel="icon" type="image/png" sizes="32x32" href="%PUBLIC_URL%/images/favicon-16x16.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="%PUBLIC_URL%/images/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/images/favicon-96x96.png">
 
     <!-- 
       Indicating where is placed the manifest.json that contains other
