+++ src/client/scripts/copyToClipboard.ts
